// src/common/modules/auth/enums/permission.enum.ts
export enum Permission {
  // Roles
  CREATE_ROLE = 'create.role',
  READ_ROLE = 'read.role',
  UPDATE_ROLE = 'update.role',
  DELETE_ROLE = 'delete.role',
  ADD_PERMISSION_ROLE = 'add.permission.role',
  REMOVE_PERMISSION_ROLE = 'remove.permission.role',
  // Users
  CREATE_USER = 'create.user',
  READ_USER = 'read.user',
  UPDATE_USER = 'update.user',
  DELETE_USER = 'delete.user',
  ADD_ROLE_USER = 'add.role.user',
  REMOVE_ROLE_USER = 'remove.role.user',
}